app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kubernetes==28.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pyyaml==6.0.1
prometheus-client==0.19.0
httpx==0.25.2
croniter==2.0.1

